move time consuming function into framework
